// src/navigation/enums.ts
export enum StackName {
  Splash = 'SplashStack',
  NonSecure = 'NonSecureStack',
  Secure = 'SecureStack',
}

export enum SplashScreenName {
  Splash = 'Splash',
}

export enum NonSecureScreenName {
  Login = 'Login',
  Register = 'Register',
}

export enum SecureScreenName {
  Tabs = 'Tabs',
}

export enum SecureTabName {
  Home = 'Home',
  Explore = 'Explore',
  Activity = 'Activity',
  Profile = 'Profile',
}